backend/requirements.txt
# Web框架
fastapi
uvicorn[standard]

# 数据库与ORM
sqlalchemy
alembic

# 异步任务队列
celery
redis

# AI与LangChain核心
langchain
langchain-openai
langchain-community
langchain-core
openai
google-search-results
duckduckgo-search

# 工具与其他
python-dotenv
requests
beautifulsoup4
lxml
readability-lxml